Common functions

IQ Mismatch

RFImpairmentsModels.addIQMismatch — Method

Apply IQ Mismatch to complex input signal x parametrized by the gain mismtach g (in dB) and the phase mismatch ϕ. The model split the gain in 2 equi-partioned gains and split the phase into equi-partioned phase. [1] Matthias Hesse, Marko Mailand, Hans-Joachim Jentschel, Luc Deneire, Jerome Lebrun. Semi-Blind Cancellation of IQ-Imbalances. IEEE International Conference on Communications, May 2008, Bei- jing, China. hal-00272886v1 There are 3 versions for the function

One with allocation of the output y = addIQMismatch(x,g,ϕ) or y = addIQMismatch(g,s::IQMismatch)

One with mutation of the second input addIQMismatch!(y,x,...)

One with in place mutation: addIQMismatch!(x,...). In the latter case the signal is modified with a signal impaired by IQ mismatch

RFImpairmentsModels.db_to_linear — Method

Convert log in linear scale

RFImpairmentsModels.deg_to_rad — Method

Convert input degree in radians
rad = degtorad(deg)

deg : Input, in degree

RFImpairmentsModels.initIQMismatch — Method

Initiate IQ mismatch model, based on gain imbalance (in dB) and phase imbalance (in radians)
initIQMismatch(g,ϕ)

g : gain imbalance, in dB

ϕ : Phase imbalance in radians

Carrier Frequency Offset

RFImpairmentsModels.addCFO — Method

Adding Carrier Frequency Offset δ in Hz to the input signal x sampled at frequency fs (In Hz) with initial phase ϕ (default 0)
This function does not mutate the input signal. See addCFO! for mutating function. In case you want to add normalized CFO, set the fs to 1.
y = addCFO(x,δ,fs,ϕ=0)
Input Parameters

x : Input signal

δ : Carrier frequency offset [Hz]

fs : Sampling rate [Hz]

ϕ : Offset phase (Default 0) [Hz]

Output parameters

y : Signal with CFO

There are 3 versions

One with allocation on the output y = addCFO(x,δ,fs,ϕ) or y = addCFO(x,cfo::CFO)

One without allocation and mutation of first input : addCFO!(y,x,...)

One with in place location of input addCFO!(x,...)

RFImpairmentsModels.initCFO — Function

Init a Carrier Frequency structure, usefull to call addCFO with scfo \ scfo = addCFO(f,fs,ϕ) Input Parameters

δ : Carrier frequency offset [Hz]

fs : Sampling rate [Hz]

ϕ : Offset phase (Default 0) [Hz]

Output parameters

s_cfo : CFO structure (CFO)

Phase Noise

RFImpairmentsModels.addPhaseNoise! — Method

Add phase noise parametrized by pn to the input signal x There are 3 versions

One with allocation of the output y = addPhaseNoise(x,pn)

One with in place mutation of the first vector addPhaseNoise!(y,x,pn)

One with in place mutation of the input vector addPhaseNoise!(x,pn)

RFImpairmentsModels.initPhaseNoise — Method

Instantiate a phase noise model, with specific model and associated parameters

Wiener => Wiener phase noise parametrized by the normalized state noise variance

None => No phase noise. Signal will not be impaired when addPhaseNoise will be used

Non linear PA

RFImpairmentsModels.addNonLinearPA — Method

Apply Power Amplifier non linearity to input signal x following non linear PA model defined in structure pa There are 3 version

One with allocation of the output y = addNonLinearPA(x,pa)

One with in place mutation of the first input addNonLinearPA!(y,x,pa)

One with in place mutation of the input signal addNonLinearPA!(x,pa)

RFImpairmentsModels.initNonLinearPA — Method

Instantiate a non linear PA model, based on a given model and is associated parameters.

Rapp PA model [:Rapp]

First model is Rapp model [1] that is function of a certain backOff power and smoothness parameter (also called the knee parameter). Rapp non linear power amplifier model is only a AM/AM distortion model and do not affect the phase. It is also possible to specify the theoretical input power value (i.e E[|x[n]|²]). Indded the saturation is defined as a backoff with respect to the input average power. It the power is not specified, the estimated power is first calculated from the time sequence (with 1/N ∑ | x[n]|²). It may lead to bias especially for signal with strong peak to average power ratio. In this case it can be usefull to compute the theoretical power first (and give to this function with power keyword) to have a constant saturation value.
[1] = C. Rapp. Effects of HPA-nonlinearity on a 4-DPSK/OFDM-signal for a digital sound broadcasting signal. In P. S. Weltevreden, editor, ESA Special Publication, volume 332 of ESA Special Publication, Oct. 1991.

Saleh Model

A second model is the Saleh model that both affect the AM/AM distortion and the AM/PM distortion. The model is parametrized by 4 parameters αAM, βAM, αPM and βPM. [2] = A. A. M. Saleh. Frequency-independent and frequency-dependent nonlinear models of TWT amplifiers. 29(11):1715–1720.

Ghorbani model

A third model is the Ghorbani model that also defines an AM/AM distortion and the AM/PM distortion through a set of 8 parameters, a1,a2,a3,a4 (for the amplitude) and b1,b2,b3,b4 for the phase distortion.

[3] = A. Ghorbani and M. Sheikhan. The effect of solid state power amplifiers (SSPAs) nonlinearities on MPSK and M-QAM signal transmission. In Digital Processing of Signals in Communications, 1991., Sixth International Conference On, pages 193–197

None model

Pure linear amplification, usefull to be sure no impairments is applied

initNonLinearPA(:Rapp;power=-1,backOff=0,smoothness=0)
- power : For Rapp model. Theoretical power for constant saturation level. If -1, empirical power is used (default -1)

backOff : For Rapp model. BackOff value, in dB (default 0)

smoothness: For Rapp model. Smoothness value (typically between 1 and 10)(default 0)

α_AM : For Saleh model. α parameter for AM distortion (default 0)

β_AM : For Saleh model. β parameter for AM distortion (default 0)

α_PM : For Saleh model. α parameter for PM distortion (default 0)

β_PM : For Saleh model. β parameter for PM distortion (default 0)

a1,a2,a3,a4 : For Ghorbani model. Parameters for AM/AM distortion (default all at 0)

b1,b2,b3,b4 : For Ghorbbni model. Pbrbmeters for AM/PM distortion (default all at 0)

linearGain : For Linear PA model. Applied gain (default 1)